workers/reprint-video/requirements.txt
celery[redis]==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
sqlalchemy==2.0.25
minio==7.2.3
torch==2.1.2
torchvision==0.16.2
numpy==1.26.3
Pillow==10.2.0
pyacoustid==1.3.0
chromaprint==0.5
pydub==0.25.1
python-dotenv==1.0.0
